+++ schema/starting-schema.sql
@@ -34,9 +34,9 @@ CREATE TABLE IF NOT EXISTS `player_gear` (
     `attack` int(11) unsigned NOT NULL DEFAULT 0,
     `defense` int(11) unsigned NOT NULL DEFAULT 0,
     `speed` int(11) unsigned NOT NULL DEFAULT 0,
-    `crit` smallint(4) unsigned NOT NULL DEFAULT 0,
-    `dodge` smallint(4) unsigned NOT NULL DEFAULT 0,
-    `block` smallint(4) unsigned NOT NULL DEFAULT 0,
+    `crit` decimal(5,2) NOT NULL DEFAULT 0,
+    `dodge` decimal(5,2) NOT NULL DEFAULT 0,
+    `block` decimal(5,2) NOT NULL DEFAULT 0,
     CONSTRAINT UC_player_slot UNIQUE (player_id, gear_slot_id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
